proiect_efficientnetv2s/requirements.txt
torch
torchvision
numpy
pillow
matplotlib
scikit-learn
tqdm
opencv-python
